+++ pkgs/code_builder/lib/code_builder.dart
@@ -17,6 +17,7 @@ export 'src/specs/enum.dart'
 export 'src/specs/expression.dart'
     show
         BinaryExpression,
+        ClosureExpression,
         CodeExpression,
         Expression,
         ExpressionEmitter,
@@ -25,6 +26,9 @@ export 'src/specs/expression.dart'
         InvokeExpressionType,
         LiteralExpression,
         LiteralListExpression,
+        LiteralMapExpression,
+        LiteralRecordExpression,
+        LiteralSetExpression,
         ParenthesizedExpression,
         ToCodeExpression,
         declareConst,
@@ -70,3 +74,4 @@ export 'src/specs/type_function.dart' show FunctionType, FunctionTypeBuilder;
 export 'src/specs/type_record.dart' show RecordType, RecordTypeBuilder;
 export 'src/specs/type_reference.dart' show TypeReference, TypeReferenceBuilder;
 export 'src/specs/typedef.dart' show TypeDef, TypeDefBuilder;
+export 'src/visitors.dart' show SpecVisitor;