+++ pkgs/code_builder/lib/src/allocator.dart
@@ -90,7 +90,6 @@ class _PrefixedAllocator implements Allocator {
   int _nextKey() => _keys++;
 
   @override
-  Iterable<Directive> get imports => _imports.keys.map(
-        (u) => Directive.import(u, as: '_i${_imports[u]}'),
-      );
+  Iterable<Directive> get imports =>
+      _imports.keys.map((u) => Directive.import(u, as: '_i${_imports[u]}'));
 }