+++ pkgs/code_builder/lib/src/specs/directive.g.dart
@@ -23,14 +23,14 @@ class _$Directive extends Directive {
   factory _$Directive([void Function(DirectiveBuilder)? updates]) =>
       (new DirectiveBuilder()..update(updates)).build() as _$Directive;
 
-  _$Directive._(
-      {this.as,
-      required this.url,
-      required this.type,
-      required this.show,
-      required this.hide,
-      required this.deferred})
-      : super._() {
+  _$Directive._({
+    this.as,
+    required this.url,
+    required this.type,
+    required this.show,
+    required this.hide,
+    required this.deferred,
+  }) : super._() {
     BuiltValueNullFieldError.checkNotNull(url, r'Directive', 'url');
     BuiltValueNullFieldError.checkNotNull(type, r'Directive', 'type');
     BuiltValueNullFieldError.checkNotNull(show, r'Directive', 'show');
@@ -189,19 +189,32 @@ class _$DirectiveBuilder extends DirectiveBuilder {
   Directive build() => _build();
 
   _$Directive _build() {
-    final _$result = _$v ??
+    final _$result =
+        _$v ??
         new _$Directive._(
-            as: as,
-            url:
-                BuiltValueNullFieldError.checkNotNull(url, r'Directive', 'url'),
-            type: BuiltValueNullFieldError.checkNotNull(
-                type, r'Directive', 'type'),
-            show: BuiltValueNullFieldError.checkNotNull(
-                show, r'Directive', 'show'),
-            hide: BuiltValueNullFieldError.checkNotNull(
-                hide, r'Directive', 'hide'),
-            deferred: BuiltValueNullFieldError.checkNotNull(
-                deferred, r'Directive', 'deferred'));
+          as: as,
+          url: BuiltValueNullFieldError.checkNotNull(url, r'Directive', 'url'),
+          type: BuiltValueNullFieldError.checkNotNull(
+            type,
+            r'Directive',
+            'type',
+          ),
+          show: BuiltValueNullFieldError.checkNotNull(
+            show,
+            r'Directive',
+            'show',
+          ),
+          hide: BuiltValueNullFieldError.checkNotNull(
+            hide,
+            r'Directive',
+            'hide',
+          ),
+          deferred: BuiltValueNullFieldError.checkNotNull(
+            deferred,
+            r'Directive',
+            'deferred',
+          ),
+        );
     replace(_$result);
     return _$result;
   }