+++ pkgs/code_builder/lib/src/specs/extension.dart
@@ -45,10 +45,7 @@ abstract class Extension extends Object
   String? get name;
 
   @override
-  R accept<R>(
-    SpecVisitor<R> visitor, [
-    R? context,
-  ]) =>
+  R accept<R>(SpecVisitor<R> visitor, [R? context]) =>
       visitor.visitExtension(this, context);
 }